Added 7-puts_half.c

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -9,7 +9,7 @@
 
 void puts_half(char *s)
 {
-	int i = 0, j, length = 0;
+	long int i = 0, j, length = 0;
 
 	while (*(s + i) != '\0')
 	{
